fix: move company components under the components/company
