+++ examples/app-crm/src/components/company/index.ts
@@ -1,0 +1,2 @@
+export * from "./card-view";
+export * from "./table-view";
